% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betacdf.R
\name{betacdf}
\alias{betacdf}
\title{distribution function for the Beta distribution}
\usage{
betacdf(a, b, x)
}
\arguments{
\item{a, b}{non-negative parameters of the Beta distribution}

\item{x}{the value of quantile}
}
\value{
the value of the Beta cdf at x
}
\description{
a function to compute a Monte Carlo estimate of the Beta cdf Using antithetic variables method
}
\examples{
\dontrun{
betacdf(3,3,0.3)
}
}
