% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betasample.R
\name{betasample}
\alias{betasample}
\title{generate samples from the Beta distribution}
\usage{
betasample(n, a, b)
}
\arguments{
\item{n}{number of observations}

\item{a, b}{non-negative parameter of the Beta distribution}
}
\value{
the random samples of size n
}
\description{
a function to generate random samples from the Beta distribution by the acceptance-rejection method
}
\examples{
\dontrun{
sample=betasample(1000,3,2)
plot(sample)
}
}
