% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rayleighsample.R
\name{Rayleighsample}
\alias{Rayleighsample}
\title{generate samples from a Rayleigh distribution}
\usage{
Rayleighsample(n, sigma, antithetic = TRUE)
}
\arguments{
\item{n}{number of observations}

\item{sigma}{non-negative parameter of the Rayleigh distribution}

\item{antithetic}{logical;if TRUE the antithetic variables method will be used}
}
\value{
the random samples of size n
}
\description{
a function to generate samples from a Rayleigh distribution
}
\examples{
\dontrun{
Rayleighsample(1000,2)
sample=Rayleighsample(1000,2,antithetic=FALSE)
plot(sample)
}
}
